Extracted test_suite_assertions
